Added Paths Endpoint
The code for implementing user_ids and startup_ids should be checked on lines 31-33 of paths.rb and lines 8-10 for file paths_spec.rb in the unit tests.  The "user_ids" and "startup_ids" are options so need to be passed in as an array to the "options" hash along with "direction."  Not sure this is done correctly on the lines mentioned.  The test passes on my local, but not sure how it will do with real data.
The implementation spec needs to be updated as there wasn't yet the ability to test with authentication.